requirements.txt
boto3==1.4.7
botocore==1.7.44
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
docker==2.6.1
docker-pycreds==0.2.1
docutils==0.14
Flask==0.12.2
google-auth==1.2.1
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kubernetes==3.0.0
MarkupSafe==1.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.11
six==1.11.0
urllib3==1.22
websocket-client==0.40.0
Werkzeug==0.12.2
Tornado==4.4.2
coralogix_logger==1.0.28
